Add JOIN Support

+++ crates/yuzu_hir/src/hir/literal.rs
@@ -1,10 +1,12 @@
+use yuzu_core::adt::{Float, Int};
+
 use crate::hir::SymbolId;
 
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub enum Literal {
     Bool { value: bool },
-    Int { value: u64 },
-    Float { value: u64 },
+    Int { value: Int },
+    Float { value: Float },
     String { value: SymbolId },
     Missing,
 }
